+++ src/state/entity/dummy_DummyPlayer.ts
@@ -28,7 +28,7 @@ function getEntityState(this: dummy.DummyPlayer, player?: StateKey) {
 
         username: memory.diff(this.username),
         skin: memory.diff(this.currentSkinName ?? ''),
-        isControlBlocked: memory.diff(this.isControlBlocked),
+        remoteForceBlockControl: memory.diff(this.remoteForceBlockControl),
         inCutscene: memory.diff(this.inCutscene),
         currentMenu: memory.diff(this.currentMenu as u32),
         currentSubState: memory.diff(this.currentSubState),
@@ -46,7 +46,7 @@ function setEntityState(this: dummy.DummyPlayer, state: Return) {
         this.setSkin(state.skin, true)
     }
 
-    if (state.isControlBlocked !== undefined) this.isControlBlocked = state.isControlBlocked
+    if (state.remoteForceBlockControl !== undefined) this.remoteForceBlockControl = state.remoteForceBlockControl
     if (state.inCutscene !== undefined) this.inCutscene = state.inCutscene
     if (state.currentMenu !== undefined) this.currentMenu = state.currentMenu
     if (state.currentSubState !== undefined) this.currentSubState = state.currentSubState