+++ src/state/entity/ig_ENTITY_Door.ts
@@ -17,13 +17,24 @@ type Return = ReturnType<typeof getEntityState>
 function getEntityState(this: ig.ENTITY.Door, player?: StateKey) {
     const memory = StateMemory.getBy(this, player)
 
+    const opened = this.lastOpened?.frame == ig.system.frame - 1
     return {
         ...igAnimatedEntity.getEntityState.call(this, memory),
+
+        opened: opened ? true : undefined,
+        openGlobalSound: memory.diff(this.lastOpened?.globalSound),
     }
 }
 
 function setEntityState(this: ig.ENTITY.Door, state: Return) {
     igAnimatedEntity.setEntityState.call(this, state)
+
+    if (state.opened === true) {
+        if (this.openSound) {
+            if (state.openGlobalSound) this.openSound.play()
+            else ig.SoundHelper.playAtEntity(this.openSound, this)
+        }
+    }
 }
 
 prestart(() => {
@@ -36,3 +47,19 @@ prestart(() => {
     }
     registerNetEntity({ entityClass: ig.ENTITY.Door, isStatic: true })
 }, 2)
+
+declare global {
+    namespace ig.ENTITY {
+        interface Door {
+            lastOpened?: { frame: number; globalSound?: boolean }
+        }
+    }
+}
+prestart(() => {
+    ig.ENTITY.Door.inject({
+        open(globalSound, openTimer) {
+            this.parent(globalSound, openTimer)
+            this.lastOpened = { frame: ig.system.frame, globalSound }
+        },
+    })
+})