+++ src/state/entity/ig_ENTITY_Combatant-base.ts
@@ -9,6 +9,7 @@ import type { f32, u16, u4 } from 'ts-binarifier/src/type-aliases'
 import { runTasks } from 'cc-instanceinator/src/inst-util'
 import { assert } from '../../misc/assert'
 import type { StateKey } from '../map-state-handlers'
+import type { EntityNetid } from '../../misc/entity-netid'
 
 declare global {
     namespace sc {
@@ -42,6 +43,8 @@ export function getEntityState(this: ig.ENTITY.Combatant, player: StateKey | und
         combatantLabelOffY: memory.diff(this.combatantLabelInfo?.offY),
 
         statusGui: memory.diffRecord(getStatusEntries(this)),
+
+        target: memory.diff(this.target?.netid ?? 0 as EntityNetid),
     }
 }
 
@@ -112,6 +115,20 @@ export function setEntityState(this: ig.ENTITY.Combatant, state: Return) {
                 )
                 ig.gui.addGuiElement(box)
             })
+        }
+    }
+
+    if (state.target !== undefined) {
+        if (state.target === 0) {
+            this.setTarget(null)
+        } else {
+            const entity = ig.game.entitiesByNetid[state.target]
+            if (!entity) {
+                console.warn('ig.ENTITY.Combatant target not found:', state.target)
+            } else {
+                assert(entity instanceof ig.ENTITY.Combatant)
+                this.setTarget(entity)
+            }
         }
     }
 }