+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Fix footstep sounds playing when dashing on remote
 - Fix player lock status not working after entering a new map
 - Fix warnings on some combat arts
+- Fix wrong health bar color on remote for combatants with target
 
 ## [0.8.3] 2026-08-01